Create README.md
Detroit 01-01-2024
Andres Rivas (amateur)
Medium electronics technician
Radiotechnician amateur

I am in the middle of research on a version of artificial intelligence dedicated to explaining how to survive in any situation. that can work remotely with private servers, and make an application that connects to that server through a 5 ghz connection, you don't have to know everything about the beginning of the world but you should know how to light a fire, types of rocks, types of soil, how to fish , how to hunt, an accurate and updatable map of the world, how to plant and raise animals for consumption and emotional care, agricultural help. If there is a storm, give an explanation why it is raining. And the benefits of rain. If you have more ideas I would love to talk to someone and exchange ideas.

Blessings.
Andres Rivas
Antruzzz
Andresrs